adding Tests\Linq to linq2db.core.sln (FW4 & FW 4.5)
