+++ Tests/Linq/Tests.project.json
@@ -1,4 +1,4 @@
 ﻿{
-	"runtimes": { "win-x86": {} },
+	"runtimes": { "win-x86": {}, "win":{} },
 	"frameworks": { "net45": {} }
 }
