+++ Tests/Linq/project.json
@@ -6,81 +6,95 @@
 
 
 	"buildOptions": {
-		"outputName": "linq2db.Tests"
+		"outputName": "linq2db.Tests",
+		"publicSign": true,
+		"keyFile": "linq2db.Tests.snk"
 	},
 
 	"dependencies": {
-		"Source": "1.7.5-*"
-
+		"Model": "1.0.0-*",
+		"NUnit": "3.4.1"
 	},
 
 	"frameworks": {
-		"net40": {
-			"frameworkAssemblies": {
-				"Microsoft.CSharp": "",
-				"System": "",
-				"System.Configuration": "",
-				"System.Core": "",
-				"System.Data.DataSetExtensions": "",
-				"System.Data.Linq": "",
-				"System.Data.Services": "",
-				"System.ServiceModel": "",
-				"System.Web.Services": "",
-				"System.Xml.Linq": "",
-				"System.Data": "",
-				"System.Xml": "",
-				"System.Runtime.Serialization": ""
-			},
-					"dependencies": {
-						"NUnit": "3.4.1",
-						"Microsoft.Bcl.Async": "1.0.*",
-						"FirebirdSql.Data.FirebirdClient": "4.10.0",
-						"Microsoft.SqlServer.Compact": "4.0.8876.1",
-						"Microsoft.SqlServer.Types": "11.0.2",
-						"MiniProfiler": "3.2.0.157",
-						"MySql.Data": "6.9.8",
-						"Npgsql": "2.2.6",
-						"Oracle.DataAccess.x86": "2.112.1.0",
-						"Oracle.ManagedDataAccess": "12.1.24160419",
-						"System.Data.SQLite.Core": "1.0.101.0",
-						"xunit": "1.9.1",
-						"xunit.extensions": "1.9.1"
+			"net40": {
+				"frameworkAssemblies": {
+					"Microsoft.CSharp": "",
+					"System": "",
+					"System.Configuration": "",
+					"System.Core": "",
+					"System.Data.DataSetExtensions": "",
+					"System.Data.Linq": "",
+					"System.Data.Services": "",
+					"System.ServiceModel": "",
+					"System.Web.Services": "",
+					"System.Xml.Linq": "",
+					"System.Data": "",
+					"System.Xml": "",
+									"System.Runtime.Serialization": "",
+									"System.Windows.Forms": ""
+				},
+				"dependencies": {
+					"Microsoft.Bcl.Async": "1.0.*",
+					"FirebirdSql.Data.FirebirdClient": "4.10.0",
+					"Microsoft.SqlServer.Compact": "4.0.8876.1",
+					"Microsoft.SqlServer.Types": "11.0.2",
+					"MiniProfiler": "3.2.0.157",
+					"MySql.Data": "6.9.8",
+					"Npgsql": "2.2.6",
+					"Oracle.DataAccess.x86": "2.112.1.0",
+					"Oracle.ManagedDataAccess": "12.1.24160419",
+					"System.Data.SQLite.Core": "1.0.101.0",
+					"xunit": "1.9.1",
+					"xunit.extensions": "1.9.1"
 
+				},
+				"buildOptions": {
+									"define": [ "FW4", "NOASYNC", "NOIMMUTABLE", "NOFSHARP", "NPGSQL226" ],
+					"compile": {
+											"exclude": [
+												"**/VisualBasicTests.cs",
+												"**/FSharpTests.cs",
+												"**/AsyncTests.cs",
+												"**/ParserTests.cs",
+												"**/DB2Tests.cs",
+												"**/InformixTests.cs"
+											]
+					}
+				}
+
+			},
+
+		"net45": {
+					"frameworkAssemblies": {
+						"Microsoft.CSharp": "",
+						"System": "",
+						"System.Configuration": "",
+						"System.Core": "",
+						"System.Data.DataSetExtensions": "",
+						"System.Data.Linq": "",
+						"System.Data.Services": "",
+						"System.ServiceModel": "",
+						"System.Web.Services": "",
+						"System.Xml.Linq": "",
+						"System.Data": "",
+						"System.Xml": "",
+						"System.Runtime.Serialization": "",
+						"System.Windows.Forms": ""
 					},
 			"buildOptions": {
-				"define": [ "FW4", "NOASYNC" ],
+							"define": [ "FW4", "NOFSHARP" ],
 				"compile": {
-					"exclude": [
-					]
-				}
-			}
-
-		},
-		"net45": {
-			"frameworkAssemblies": {
-				"Microsoft.CSharp": "",
-				"System": "",
-				"System.Configuration": "",
-				"System.Core": "",
-				"System.Data.DataSetExtensions": "",
-				"System.Data.Linq": "",
-				"System.Data.Services": "",
-				"System.ServiceModel": "",
-				"System.Web.Services": "",
-				"System.Xml.Linq": "",
-				"System.Data": "",
-				"System.Xml": "",
-				"System.Runtime.Serialization": ""
-			},
-			"buildOptions": {
-				"define": [ "FW4" ],
-				"compile": {
-					"exclude": [
-					]
+									"exclude": [
+										"**/VisualBasicTests.cs",
+										"**/FSharpTests.cs",
+										"**/ParserTests.cs",
+										"**/DB2Tests.cs",
+										"**/InformixTests.cs"
+									]
 				}
 			},
 			"dependencies": {
-				"NUnit": "3.4.1",
 
 				"FirebirdSql.Data.FirebirdClient": "4.10.0",
 				"Microsoft.Bcl.Immutable": "1.0.34",
@@ -88,7 +102,7 @@
 				"Microsoft.SqlServer.Types": "11.0.2",
 				"MiniProfiler": "3.2.0.157",
 				"MySql.Data": "6.9.8",
-				"Npgsql": "2.2.6",
+				"Npgsql": "3.1.8",
 				"Oracle.DataAccess.x86": "2.112.1.0",
 				"Oracle.ManagedDataAccess": "12.1.24160419",
 				"System.Data.SQLite.Core": "1.0.101.0",
