+++ linq2db.core.sln
@@ -19,6 +19,8 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Tests.NetCore", "Tests\Test
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Model", "Tests\Model\Model.xproj", "{EF534F71-A4E6-4019-8835-43BEE792C70A}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Linq", "Tests\Linq\Linq.xproj", "{471EE14E-D11F-4E55-9293-7E9617573168}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -37,6 +39,10 @@ Global
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF534F71-A4E6-4019-8835-43BEE792C70A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{471EE14E-D11F-4E55-9293-7E9617573168}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{471EE14E-D11F-4E55-9293-7E9617573168}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{471EE14E-D11F-4E55-9293-7E9617573168}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{471EE14E-D11F-4E55-9293-7E9617573168}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -45,5 +51,6 @@ Global
 		{0E4E9A56-F285-4067-ADAA-392ACB9CAAA7} = {AF93451E-198C-4A94-8EB4-C330D886E102}
 		{F84DD030-1FCB-4D7C-88CE-62B1BFB078D5} = {DF8DE665-05A2-4ACF-9799-8FC35E7F1525}
 		{EF534F71-A4E6-4019-8835-43BEE792C70A} = {DF8DE665-05A2-4ACF-9799-8FC35E7F1525}
+		{471EE14E-D11F-4E55-9293-7E9617573168} = {DF8DE665-05A2-4ACF-9799-8FC35E7F1525}
 	EndGlobalSection
 EndGlobal
